Update finalproject.Rmd
Connors, A.F. et al. (1996). The effectiveness of right heart catheterization in the initial care of critically III patients. JAMA, 276(11), 889-897. 
The objective of this study is to examine the association between the use of right heart catheterization (RHC) during the first 24 hours of care in the intensive care unit (ICU) and sub- sequent survival, length of stay, intensity of care, and cost of care. This dataset is available at http://biostat.mc.vanderbilt.edu/wiki/Main/DataSets under “Right heart catheterization dataset ”. Your group is in charge of the following analyses: 
1.	Perform all relevant descriptive analyses. 
2.	Build a model to predict whether a patient will receive RHC as a function of all pre-RHC variables. Comment on the interpretation of the coefficients of this model. Comment on potential issues regarding confounding of the effect of RHC on clinical outcomes in this study. 
3.	Build a model to predict whether a patient will die in the first 180 days after hospitalization, based on all pre-treatment covariates. Comment on the interpretation of the coefficients of this model. What is the marginal risk ratio of death associated to RHC? Compute a 95% confidence interval. Comment on the possibility of interpreting this estimate as a causal effect. 
4.	A co-investigator in this study tells you that she believes there is a differential effect of RHC on 180-day mortality according to primary disease category. Is there any evidence in the data to support this claim? Comment.