Create DNA sequencing and Cardiac Arrythmia
Background: The international HapMap project (Links to an external site.)Links to an external site. made possible the elucidation of the entire human genome. The haplotype map, or "HapMap," is a tool that allows researchers to find genes and genetic variations that affect health and disease. The DNA sequence of any two people is 99.5 percent identical. The variations, however, may greatly affect an individual's disease risk. Sites in the DNA sequence where individuals differ at a single DNA base are called single nucleotide polymorphisms (SNPs). Sets of nearby SNPs on the same chromosome are inherited in blocks. This pattern of SNPs on a block is a haplotype. Blocks may contain a large number of SNPs, but a few SNPs are enough to uniquely identify the haplotypes in a block. The HapMap is a map of these haplotype blocks and the specific SNPs that identify the haplotypes are called tag SNPs.

Data: Simulated genotypes of 100,000 markers are provided for two populations: CHB or Han Chinese in Beijing, China (N=137) and JPT or Japanese in Tokyo, Japan (N=113). The genotypes at these markers are coded as 0, 1 and 2 for the number of counts of the SNP (maximum possible is 2). 

Question 1 a) Calculate a similarity measure between individuals (called the Jaccard matrix) i.e. if there are N individuals create a NxN matrix of similarities. The Jaccard measure of similarity between two individuals A and B is defined as 

Question 1b) Using the Jaccard similarity matrix perform a principal component analysis. How many PCs explain most of the variation in the data? Plot the PC scores for the first two PC components and label the points with different colors denoting the two populations CHB and JPT.

Question 2. (75 points)

The aim is to distinguish between the presence and absence of cardiac arrhythmia and to classify it in one of the 16 groups based on 12-lead ECG (or Electrocardiography) measurements on 452 individuals. A computer program makes such a classification but there are differences between the cardiologist's and the program's classification. Taking the cardiologist as a gold standard the goal is to minimize this difference by means of machine learning tools. List ALL methods that you have learnt in this course which can be applied to this problem. Compare the performance of all (or some) of these methods in predicting the presence or absence of cardiac arrhythmia. If you choose to drop a few methods from your comparative study of various methods you must provide a clear rationale to do so. Appropriate methods to estimate the error rate must be employed and the rationale for choosing the method must be described clearly. Finally, choose the best performing method and compare with the results obtained in the original research article (see below).